requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
python-slugify==8.0.1
selenium==4.16.0

